+++ src/localllm/models.py
@@ -19,6 +19,7 @@ def localllm_download_model(
                 
         - "gemma-3-270m-it-Q8_0": Google Gemma 3 270M it model (Q8_0 quantization)
         - "gemma-3-270m-it-qat-Q4_0": Google Gemma 3 270M it model (Q4_0 quantization)
+        - "gemma-3-4b-it-Q4_K_M": Google Gemma 3 4b it model (Q4_K_M quantization)
         - "GLM-4.6V-Flash-Q4_K_M": GLM 4.6V Flash model (Q4_K_M quantization)
         - "translategemma-4b-it-q8_0": TranslateGemma 4B it model (Q8_0 quantization)
         
@@ -53,6 +54,8 @@ def localllm_download_model(
     >>> os.remove(model_path)  # Clean up downloaded file
     >>> model_path = localllm_download_model("gemma-3-270m-it-qat-Q4_0", overwrite=True)
     Downloading...
+    >>> model_path = localllm_download_model("gemma-3-4b-it-Q4_K_M", overwrite=True)
+    Downloading...
     >>> model_path = localllm_download_model("GLM-4.6V-Flash-Q4_K_M", overwrite=True)
     Downloading...
     >>> model_path = localllm_download_model("translategemma-4b-it-q8_0", overwrite=True)
@@ -68,20 +71,24 @@ def localllm_download_model(
     model_registry = {
         "gemma-3-270m-it-Q8_0": {
             "url": "https://huggingface.co/bartowski/google_gemma-3-270m-it-GGUF/resolve/main/google_gemma-3-270m-it-Q8_0.gguf",
-            "filename": "google_gemma-3-270m-it-Q8_0.gguf"
+            "filename": "google_gemma-3-270m-it-Q8_0.gguf",
         },
         "gemma-3-270m-it-qat-Q4_0": {
             "url": "https://huggingface.co/ggml-org/gemma-3-270m-it-qat-GGUF/resolve/main/gemma-3-270m-it-qat-Q4_0.gguf",
-            "filename": "gemma-3-270m-it-qat-Q4_0.gguf"
+            "filename": "gemma-3-270m-it-qat-Q4_0.gguf",
+        },
+        "gemma-3-4b-it-Q4_K_M": {
+            "url": "https://huggingface.co/bartowski/google_gemma-3-4b-it-GGUF/resolve/main/google_gemma-3-4b-it-Q4_K_M.gguf",
+            "filename": "google_gemma-3-4b-it-Q4_K_M.gguf",
         },
         "GLM-4.6V-Flash-Q4_K_M": {
             "url": "https://huggingface.co/ggml-org/GLM-4.6V-Flash-GGUF/resolve/main/GLM-4.6V-Flash-Q4_K_M.gguf",
-            "filename": "GLM-4.6V-Flash-Q4_K_M.gguf"
+            "filename": "GLM-4.6V-Flash-Q4_K_M.gguf",
         },
         "translategemma-4b-it-q8_0": {
             "url": "https://huggingface.co/NikolayKozloff/translategemma-4b-it-Q8_0-GGUF/resolve/main/translategemma-4b-it-q8_0.gguf",
-            "filename": "translategemma-4b-it-q8_0.gguf"
-        },        
+            "filename": "translategemma-4b-it-q8_0.gguf",
+        },
     }
     
     # Validate model type