+++ CHANGELOG.md
@@ -7,5 +7,6 @@
 * localllm_download_model:
     * gemma-3-270m-it-Q8_0
     * gemma-3-270m-it-qat-Q4_0
+    * gemma-3-4b-it-Q4_K_M
     * GLM-4.6V-Flash-Q4_K_M
     * translategemma-4b-it-q8_0